% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoseResponseModels.R
\name{betaMod}
\alias{betaMod}
\title{Implementation of Beta models}
\usage{
betaMod(d, e, scal)
}
\arguments{
\item{d}{real-valued argument to the function (dose variable)}

\item{e}{model parameter}

\item{scal}{fixed dose scaling parameter}
}
\value{
Response value.
}
\description{
Beta model: \deqn{m(d,\theta)=E_0+E_{max}B(\delta_1,\delta_2)(d/scal)^{\delta_1}(1-d/scal)^{\delta_2}}
with \deqn{B(\delta_1,\delta_2)=(\delta_1+\delta_2)^{\delta_1+\delta_2}/(\delta_1^{\delta_1} \delta_2^{\delta_2})}
and \eqn{scal} is a fixed dose scaling parameter.
}

