% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dff.R
\name{dff}
\alias{dff}
\title{Implementation of absolute difference function}
\usage{
dff(d, beta1, beta2, m1, m2)
}
\arguments{
\item{d}{real-valued argument to the function (dose variable)}

\item{beta1, beta2}{model parameters (real vectors)}

\item{m1, m2}{model types. Built-in models are "linlog",  "linear",  "quadratic",  "emax",  "exponential",  "sigEmax",  "betaMod" and "logistic"}
}
\value{
Response value for the absolute difference of two models.
}
\description{
Function calculating the absolute difference of two dose response models: \deqn{dff(d,\beta_1,\beta_2)=|m_1(d,\beta_1)-m_2(d,\beta_2)|}
}
