% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseResponseModels.R
\name{exponential}
\alias{exponential}
\title{Implementation of exponential models}
\usage{
exponential(d, e)
}
\arguments{
\item{d}{real-valued argument to the function (dose variable)}

\item{e}{model parameter}
}
\value{
Response value.
}
\description{
Exponential model: \deqn{m(d,\beta)=E_0+E_1(exp(d/\delta)-1)}
}
