% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseResponseModels.R
\name{linlog}
\alias{linlog}
\title{Implementation of linear in log models}
\usage{
linlog(d, e, off)
}
\arguments{
\item{d}{real-valued argument to the function (dose variable)}

\item{e}{model parameter}

\item{off}{fixed offset parameter}
}
\value{
Response value.
}
\description{
Linear in log Model model: \deqn{m(d,\beta)=E_0+\delta\ log(d+off)}
and \eqn{off} is a fixed offset parameter.
}
