% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseResponseModels.R
\name{logistic}
\alias{logistic}
\title{Implementation of logistic models}
\usage{
logistic(d, e)
}
\arguments{
\item{d}{real-valued argument to the function (dose variable)}

\item{e}{model parameter}
}
\value{
Response value.
}
\description{
Logistic model: \deqn{m(d,\beta)=E_0+\frac{E_{max}}{1+exp[(ED_{50}-d)/\delta]}}
}
