% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoseResponseModels.R
\name{sigEmax}
\alias{sigEmax}
\title{Implementation of Sigmoid Emax models}
\usage{
sigEmax(d, e)
}
\arguments{
\item{d}{real-valued argument to the function (dose variable)}

\item{e}{model parameter}
}
\value{
Response value
}
\description{
Sigmoid Emax Model model: \deqn{m(d,\theta)=E_0+E_{max} \frac{d^h}{ED_{50}^h+d^h}}
}

