% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dff.R
\name{dff}
\alias{dff}
\title{Implementation of absolute difference function}
\usage{
dff(d, alpha, beta, m1, m2)
}
\arguments{
\item{d}{real-valued argument to the function (dose variable)}

\item{alpha,beta}{model parameters (real vectors)}

\item{m1,m2}{model types. Built-in models are "linlog",  "linear",  "quadratic",  "emax",  "exponential",  "sigEmax",  "betaMod" and "logistic"}
}
\value{
Response value for the absolute difference of two models.
}
\description{
Function calculating the absolute difference of two dose response models: \deqn{dff(d,\theta_1,\theta_2)=|m_1(d,\theta_1)-m_2(d,\theta_2)|}
}

