% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseResponseModels.R
\name{linear}
\alias{linear}
\title{Implementation of linear models}
\usage{
linear(d, e)
}
\arguments{
\item{d}{real-valued argument to the function (dose variable)}

\item{e}{model parameter}
}
\value{
Response value.
}
\description{
Linear model: \deqn{m(d,\beta)=E_0+\delta d}
}
