% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoseResponseModels.R
\name{quadratic}
\alias{quadratic}
\title{Implementation of quadratic models}
\usage{
quadratic(d, e)
}
\arguments{
\item{d}{real-valued argument to the function (dose variable)}

\item{e}{model parameter}
}
\value{
Response value.
}
\description{
Quadratic model: \deqn{m(d,\theta)=E_0+\beta_1 d+\beta_2 d^2}
}

